test: refactor url control tests
